modernize project: tests now don't mutate original package.json move tasks files to esm as they used from cli only better esbuild watch task on f5 add editorconfig so tabs are preserved
